requirements.txt
beautifulsoup4==4.4.0
boto==2.38.0
Django==1.8.4
django-appconf==1.0.1
django-cors-headers==1.1.0
django-froala-editor==1.2.7
django-geoposition==0.2.2
django-imagekit==3.2.7
django-storages==1.1.8
pilkit==1.1.13
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.0
pytz==2015.6
six==1.9.0
wheel==0.24.0
